add Home page for prototype2 and move old routes to /feedback
